Ignore indels in the overlap accuracy calculation
Modify the overlap accuracy calculation to ignore indels, which
comprise nearly all CCS reads errors.  Report instead the "match
accuracy" = mismatches / overlapLength.  Downstream thresholds
in CCS assemblies can be much more stringent with match accuracy
to resolve repeats and segmental duplications.

DALIGNER stores the number of diffs in an overlap, which is the
sum of indels and mismatches.  To get the count of mismatches
requires recreated the full trace.

Add -I flag to ignore indels when computing accuracy